pyproject.toml
[project]
name = "personal-reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=6.3.0",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.44.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
